study/django/0415/prep/requirements.txt
asgiref==3.8.1
autopep8==2.3.1
Django==4.2.16
djangorestframework==3.15.2
pycodestyle==2.12.1
pytz==2024.1
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
tzdata==2024.2